🧐 Clarify we allow OIDC Distributed claims (#30)
Currently, we find that the the tdf_claims' `entitlements` list often exceeds the reasonable size requirements of an HTTP header. Putting it in a distributed claim frees up the space, at the cost of more complexity on the server side (during token generation and verification)

Since the entitlements are mostly state invariant, we also moved the remaining field, `client_public_signing_key`, to be a direct claim. While we do wish to remove it in favor of DPoP, it looks like DPoP is only supported on web app flows, and not client secret flows, so this field may hang around a bit longer